fix: split labels in kanban board
