+++ web/core/components/issues/issue-layouts/properties/index.ts
@@ -1,2 +1,3 @@
 export * from "./labels";
 export * from "./all-properties";
+export * from "./label-dropdown";
