+++ web/core/components/issues/issue-layouts/spreadsheet/columns/label-column.tsx
@@ -37,6 +37,7 @@ export const SpreadsheetLabelColumn: React.FC<Props> = observer((props: Props) =
         onClose={onClose}
         noLabelBorder
         fullWidth
+        fullHeight
       />
     </div>
   );
